Add RuntimeVirtioSerial
Previous approach works, but has one blocking drawback: in order to
deserialize the runtime state, and access the PtyName field, the client
application has to link to Code-Hex/vz which is a big no, as this
prevents easy cross compilation.

This tries a different approach, with separate VirtioSerial and
VirtioSerialRuntime structs. Let's see how this goes.

Signed-off-by: Christophe Fergeau <cfergeau@redhat.com>